feat(next.config): add support for GitHub Pages deployment
This change adds support for deploying the Next.js application to GitHub
Pages. It introduces a new `githubPagesNextConfig` object that sets the
`output` option to `"export"` and the `assetPrefix` to
`"/chelajs-website/"`. The main `nextConfig` object is then merged with
the `githubPagesNextConfig` if the `DEPLOY_ENV` environment variable is
set to `"github-pages"`.